Реализовал видеокарту

+++ db/Classes/VideoCard.cs
@@ -22,7 +22,7 @@ namespace db.Classes
         public string typeMemory { get; set; }
 
         [DisplayName("Частота памяти")]
-        public float frequencyMemory { get; set; }
+        public int frequencyMemory { get; set; }
 
         [DisplayName("Мощность")]
         public int power { get; set; }
@@ -38,7 +38,7 @@ namespace db.Classes
             id = Guid.NewGuid().ToString();
         }
 
-        public VideoCard(string gCPU, int volumeMemory, string typeMemory, float frequencyMemory, int power, string title, string manufacturer)
+        public VideoCard(string gCPU, int volumeMemory, string typeMemory, int frequencyMemory, int power, string title, string manufacturer)
         {
             id = Guid.NewGuid().ToString();
             GCPU = gCPU;
